ion: isolate msm-specific ion extensions
This is another step in the process of isolating msm-specific ion
features from stock ion.


Conflicts:

	include/linux/ion.h

+++ arch/arm/mach-msm/board-8064.c
@@ -502,7 +502,7 @@ static void __init reserve_ion_memory(void)
 			int fixed_position = NOT_FIXED;
 			struct ion_cp_heap_pdata *pdata = NULL;
 
-			switch (heap->type) {
+			switch ((int)heap->type) {
 			case ION_HEAP_TYPE_CP:
 				pdata =
 				(struct ion_cp_heap_pdata *)heap->extra_data;
